Economic Freedom of the World (EFW) Panel Data

Provides lightly cleaned data from the Economic Freedom of the World (EFW) reports by Gwartney et al. (2021) and published by the Fraser Institute.

data("efwpnl")

Format

A data frame with 4290 observations on the following 12 variables.

iso_code

ISO Alpha-3 Code

country_name

Country Name

year

Data Year

region_code

World Bank Region Code

region_name

World Bank Region Name

income_group

World Bank Income Group

efw_index

EFW Composite Index

size_of_govt

Area 1: Size of Government

legal_system

Area 2: Rule of Law

sound_money

Area 3: Sound Money

trade_freedom

Area 4: Freedom to Trade Internationally

regulation

Area 5: Regulation

Details

Source

References

James Gwartney, Robert Lawson, Joshua Hall, and Ryan Murphy (2021). Economic Freedom Dataset, published in Economic Freedom of the World: 2021 Annual Report. Fraser Institute.

Examples

data(efwpnl)
summary(efwpnl)

#>    iso_code         country_name            year      region_code       
#>  Length:4290        Length:4290        Min.   :1970   Length:4290       
#>  Class :character   Class :character   1st Qu.:2000   Class :character  
#>  Mode  :character   Mode  :character   Median :2006   Mode  :character  
#>                                        Mean   :2003                     
#>                                        3rd Qu.:2013                     
#>                                        Max.   :2019                     
#>                                                                         
#>  region_name        income_group         efw_index      size_of_govt  
#>  Length:4290        Length:4290        Min.   :2.349   Min.   :1.193  
#>  Class :character   Class :character   1st Qu.:5.866   1st Qu.:5.733  
#>  Mode  :character   Mode  :character   Median :6.747   Median :6.644  
#>                                        Mean   :6.605   Mean   :6.579  
#>                                        3rd Qu.:7.491   3rd Qu.:7.463  
#>                                        Max.   :9.109   Max.   :9.559  
#>                                        NA's   :723     NA's   :724    
#>   legal_system    sound_money    trade_freedom       regulation   
#>  Min.   :1.698   Min.   :0.000   Min.   : 0.0068   Min.   :1.143  
#>  1st Qu.:4.009   1st Qu.:6.715   1st Qu.: 5.8846   1st Qu.:6.003  
#>  Median :5.061   Median :8.078   Median : 7.0440   Median :6.877  
#>  Mean   :5.238   Mean   :7.752   Mean   : 6.7925   Mean   :6.724  
#>  3rd Qu.:6.322   3rd Qu.:9.303   3rd Qu.: 8.1037   3rd Qu.:7.593  
#>  Max.   :9.091   Max.   :9.922   Max.   :10.0000   Max.   :9.429  
#>  NA's   :720     NA's   :721     NA's   :796       NA's   :742